+++ NEWS.md
@@ -1,5 +1,6 @@
 # goshawk 0.1.5.9000
 
+* Descrease width of error bars in `g_lineplot`.
 * issues with no news:
 
 # goshawk 0.1.5