+++ NEWS.md
@@ -1,5 +1,6 @@
 # goshawk 0.1.5.9000
 
+* Adding additional validation and description for the `loq_legend` argument.
 * Decrease width of error bars in `g_lineplot`.
 * Add option to add rug plot in `g_density_distribution_plot`.